# api.swagger.yaml
swagger: "2.0"
info:
  title: Blog Server.
  description: Entry point for the application.
  version: 1.0.0
  contact:
    name: HardDie
    url: https://github.com/HardDie
    email: oleg1995sysoev@yandex.ru
...
basePath: /
schemes:
  - https
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/v1/invites/generate:
    post:
      summary: Generate a new invitation code
      operationId: Invite_Generate
      responses:
        "200":
          description: A successful response.
          schema:
            $ref: '#/definitions/gatewayGenerateResponse'
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      tags:
        - Invite
  /api/v1/invites/revoke:
    delete:
      summary: Revoke the generated invitation code
      operationId: Invite_Revoke
      responses:
        "200":
          description: A successful response.
          schema:
            type: object
            properties: {}
        default:
          description: An unexpected error response.
          schema:
            $ref: '#/definitions/rpcStatus'
      tags:
        - Invite
...
definitions:
...
  gatewayGenerateResponse:
    type: object
    properties:
      data:
        type: string
...
  protobufAny:
    type: object
    properties:
      '@type':
        type: string
    additionalProperties: {}
  rpcStatus:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      details:
        type: array
        items:
          type: object
          $ref: '#/definitions/protobufAny'
securityDefinitions:
  bearer:
    type: apiKey
    description: 'Authentication token, prefixed by Bearer: Bearer <token>'
    name: Authorization
    in: header
security:
  - bearer: []